+++ packages/patternfly-4/react-integration/demo-app-ts/src/components/demos/LoginPageDemo/LoginPageDemo.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+/* eslint-disable camelcase */
 const brandImg = require('./images/brandImgColor.svg') as string;
 const pfbg_1200 = require('./images/pfbg_1200.jpg');
 const pfbg_768 = require('./images/pfbg_768.jpg');
@@ -14,7 +15,7 @@ import {
   LoginPage,
   BackgroundImageSrc,
   ListItem,
-  ListVariant,
+  ListVariant
 } from '@patternfly/react-core';
 import ExclamationCircleIcon from '@patternfly/react-icons/dist/js/icons/exclamation-circle-icon';