+++ packages/patternfly-4/react-core/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.140.1](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.140.0...@patternfly/react-core@3.140.1) (2020-02-14)
+
+**Note:** Version bump only for package @patternfly/react-core
+
+
+
+
+
+# [3.140.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.139.0...@patternfly/react-core@3.140.0) (2020-02-14)
+
+
+### Features
+
+* **inline-edit:** add row level inline editing feature to Table ([#3581](https://github.com/patternfly/patternfly-react/issues/3581)) ([498d49f](https://github.com/patternfly/patternfly-react/commit/498d49fcc5cc9188197ddc7cd6468907fdd8224d))
+
+
+
+
+
+# [3.139.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.138.0...@patternfly/react-core@3.139.0) (2020-02-14)
+
+
+### Features
+
+* **SimpleList:** add simple list ([#3645](https://github.com/patternfly/patternfly-react/issues/3645)) ([4bba16c](https://github.com/patternfly/patternfly-react/commit/4bba16c3fbe788a74523ddede45293eb48a60009))
+
+
+
+
+
 # [3.138.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@3.137.0...@patternfly/react-core@3.138.0) (2020-02-13)