+++ packages/patternfly-4/react-docs/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.20.1](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.20.0...@patternfly/react-docs@4.20.1) (2020-02-14)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+# [4.20.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.19.4...@patternfly/react-docs@4.20.0) (2020-02-14)
+
+
+### Features
+
+* **inline-edit:** add row level inline editing feature to Table ([#3581](https://github.com/patternfly/patternfly-react/issues/3581)) ([498d49f](https://github.com/patternfly/patternfly-react/commit/498d49fcc5cc9188197ddc7cd6468907fdd8224d))
+
+
+
+
+
+## [4.19.4](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.19.3...@patternfly/react-docs@4.19.4) (2020-02-14)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.19.3](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.19.2...@patternfly/react-docs@4.19.3) (2020-02-14)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
 ## [4.19.2](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.19.1...@patternfly/react-docs@4.19.2) (2020-02-13)
 
 **Note:** Version bump only for package @patternfly/react-docs